Move domain error to shared domain
