+++ src/Context/Video/Module/Notification/Domain/NotificationNotFound.php
@@ -2,7 +2,7 @@
 
 namespace CodelyTv\Context\Video\Module\Notification\Domain;
 
-use CodelyTv\Exception\DomainError;
+use CodelyTv\Shared\Domain\DomainError;
 
 final class NotificationNotFound extends DomainError
 {
