+++ src/Context/Video/Module/User/Domain/UserNotExist.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace CodelyTv\Context\Video\Module\User\Domain;
 
-use CodelyTv\Exception\DomainError;
+use CodelyTv\Shared\Domain\DomainError;
 
 final class UserNotExist extends DomainError
 {
