+++ src/Context/Video/Module/Video/Domain/VideoNotFound.php
@@ -2,7 +2,7 @@
 
 namespace CodelyTv\Context\Video\Module\Video\Domain;
 
-use CodelyTv\Exception\DomainError;
+use CodelyTv\Shared\Domain\DomainError;
 
 final class VideoNotFound extends DomainError
 {
