+++ src/Shared/Contract/InvalidSecondsInterval.php
@@ -2,7 +2,7 @@
 
 namespace CodelyTv\Shared\Contract;
 
-use CodelyTv\Exception\DomainError;
+use CodelyTv\Shared\Domain\DomainError;
 use CodelyTv\Shared\Domain\Second;
 
 final class InvalidSecondsInterval extends DomainError
